requirements.txt
streamlit==1.25.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
rank-bm25==0.2.2
pymupdf==1.21.1
numpy==1.24.0
